fix:  2d bbox coordinates and refactor projection methods to BBox3D
